+++ src/tests-core/tests/prototype/btss/btss_test_factory.hpp
@@ -139,7 +139,7 @@ public:
 
 
 template <PackedSizeType LeafSizeType, PackedSizeType BranchSizeType>
-struct TypeHash<BTSSTestCtr<LeafSizeType, BranchSizeType>>:   UInt32Value<
+struct TypeHash<BTSSTestCtr<LeafSizeType, BranchSizeType>>:   UInt64Value<
     HashHelper<3011, (int32_t)LeafSizeType, (int32_t)BranchSizeType>::Value
 > {};
